Update Source Code Files
- Seperate Files : index.js (for start server and main webhooks handle) / data.js (for fetch data from api and then passing to parload) / payload.js (for store payload json object and set response value to data in payload)

- Refactor Code (A Little Bit)

- Remove Unused Var / Functions
