+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "echo-bot",
-  "version": "0.0.0",
-  "description": "An example LINE bot just to echo messages",
-  "main": "index.js",
+  "name": "Line-Chatbot-COVID-19-TH-Today",
+  "version": "2.0.0",
+  "description": "",
+  "main": "./public/index.js",
   "scripts": {
     "start": "node ."
   },
